+++ packages/terra-form-select/src/_Menu.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import { polyfill } from 'react-lifecycles-compat';
+import { injectIntl, intlShape } from 'react-intl';
 import KeyCode from 'keycode-js';
 import Variants from './_constants';
 import AddOption from './_AddOption';
@@ -9,6 +10,7 @@ import ClearOption from './_ClearOption';
 import MaxSelection from './_MaxSelection';
 import NoResults from './_NoResults';
 import Util from './_MenuUtil';
+import SharedUtil from './_SharedUtil';
 import styles from './_Menu.module.scss';
 
 const cx = classNames.bind(styles);
@@ -20,6 +22,19 @@ const propTypes = {
    * The content of the menu.
    */
   children: PropTypes.node,
+  /**
+   * Select element ref.
+   */
+  select: PropTypes.instanceOf(Element),
+  /**
+   * @private
+   * The intl object to be injected for translations.
+   */
+  intl: intlShape.isRequired,
+  /**
+   * Input element ref used in select component.
+   */
+  input: PropTypes.instanceOf(Element),
   /**
    * Text for the clear option.
    */
@@ -71,12 +86,15 @@ const propTypes = {
 
 const defaultProps = {
   children: undefined,
+  input: undefined,
   clearOptionDisplay: undefined,
   maxSelectionCount: undefined,
   noResultContent: undefined,
   onDeselect: undefined,
   optionFilter: undefined,
   searchValue: undefined,
+  select: undefined,
+  visuallyHiddenComponent: undefined,
   value: undefined,
 };
 
@@ -153,20 +171,57 @@ class Menu extends React.Component {
 
   componentDidMount() {
     document.addEventListener('keydown', this.handleKeyDown);
+    /**
+     * Without this detection for ontouchstart and the early return, VoiceOver on iOS will read the
+     * first option twice when the menu is opened. First due to aria-live update in componentDidMount
+     * and another time because we shift focus to the dropdown and VoiceOver on iOS will read the
+     * first option in the dropdown. To mitigate this, the following conditional
+     * check opts-out of the aria-live update if browser supports ontouchstart which iOS supports.
+     */
+    if ('ontouchstart' in window) return;
+    this.updateCurrentActiveScreenReader();
   }
 
   componentDidUpdate() {
-    this.updateLiveRegion();
+    this.updateNoResultsScreenReader();
     this.scrollIntoView();
   }
 
   componentWillUnmount() {
     this.clearSearch();
     this.clearScrollTimeout();
+    if (this.state.closedViaKeyEvent) {
+      if (this.props.variant === Variants.DEFAULT) {
+        this.props.select.focus();
+        if (SharedUtil.isSafari()) {
+          /**
+           * Shifting focus back to the select specifically
+           * when VoiceOver is on will sometimes trigger VO to shift focus
+           * randomly to the root of document or the skip to main link
+           * instead of the select and then break VoiceOver usage when navigating the
+           * select options on subsequent opening of select
+           * Refocusing on select seems to seems to mitigate this VO bug.
+           */
+          setTimeout(() => {
+            this.props.select.focus();
+          }, 300);
+        }
+      } else {
+        this.props.input.focus();
+      }
+    }
     document.removeEventListener('keydown', this.handleKeyDown);
   }
 
-  updateLiveRegion() {
+  isActiveSelected() {
+    if (Array.isArray(this.props.value)) {
+      return this.props.value.includes(this.state.active);
+    }
+
+    return this.state.active === this.props.value;
+  }
+
+  updateNoResultsScreenReader() {
     if (this.liveRegionTimeOut) {
       clearTimeout(this.liveRegionTimeOut);
     }
@@ -178,27 +233,73 @@ class Menu extends React.Component {
       } = this.state;
 
       const {
+        intl,
         visuallyHiddenComponent,
         searchValue,
       } = this.props;
-
-      const {
-        intl,
-      } = this.context;
 
       // Race condition can occur between calling timeout and unmounting this component.
       if (!visuallyHiddenComponent || !visuallyHiddenComponent.current) {
         return;
       }
 
-      if (hasNoResults && searchValue) {
+      if (hasNoResults) {
         visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.noResults' }, { text: searchValue });
       } else if (hasMaxSelection) {
         visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.maxSelectionOption' }, { text: this.props.maxSelectionCount });
       } else {
         visuallyHiddenComponent.current.innerText = '';
       }
-    }, 500);
+    }, 1000);
+  }
+
+  updateCurrentActiveScreenReader() {
+    const {
+      intl,
+      variant,
+      clearOptionDisplay,
+      visuallyHiddenComponent,
+    } = this.props;
+
+    const clearSelectTxt = intl.formatMessage({ id: 'Terra.form.select.clearSelect' });
+
+    if (this.menu !== null) {
+      this.menu.setAttribute('aria-activedescendant', `terra-select-option-${this.state.active}`);
+    }
+
+    // Announces options to screen readers as user navigates through them via keyboard
+    if (visuallyHiddenComponent) {
+      // Detects if option is clear option and provides accessible text
+      if (clearOptionDisplay) {
+        const active = this.menu.querySelector('[data-select-active]');
+        if (active.hasAttribute('data-terra-select-clear-option')) {
+          this.props.visuallyHiddenComponent.current.innerText = clearSelectTxt;
+        }
+      }
+
+      // Detects if option is an "Add option" and provides accessible text
+      if (variant === Variants.COMBOBOX || variant === Variants.TAG) {
+        const active = this.menu.querySelector('[data-select-active]');
+
+        if (active.hasAttribute('data-terra-select-add-option')) {
+          const display = active.querySelector("[class*='display']").innerText;
+          this.props.visuallyHiddenComponent.current.innerText = display;
+        }
+      }
+
+      const element = Util.findByValue(this.props.children, this.state.active);
+      if (element) {
+        if (element.props.display === '' && element.props.value === '') {
+          // Used for case where users selects clear option and opens
+          // dropdown again and navigates to clear option
+          this.props.visuallyHiddenComponent.current.innerText = clearSelectTxt;
+        } else if (this.isActiveSelected()) {
+          this.props.visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.selectedText' }, { text: element.props.display });
+        } else {
+          this.props.visuallyHiddenComponent.current.innerText = element.props.display;
+        }
+      }
+    }
   }
 
   /**
@@ -231,6 +332,7 @@ class Menu extends React.Component {
           isActive: option.props.value === this.state.active,
           isCheckable: Util.allowsMultipleSelections(this.props.variant),
           isSelected: Util.isSelected(this.props.value, option.props.value),
+          variant: this.props.variant,
           onMouseDown: () => { this.downOption = option; },
           onMouseUp: event => this.handleOptionClick(event, option),
           onMouseEnter: event => this.handleMouseEnter(event, option),
@@ -250,20 +352,77 @@ class Menu extends React.Component {
   handleKeyDown(event) {
     const { keyCode } = event;
     const { active, children } = this.state;
-    const { onSelect, value, variant } = this.props;
+    const {
+      intl,
+      onSelect,
+      onDeselect,
+      value,
+      variant,
+    } = this.props;
+
+    const selectedTxt = intl.formatMessage({ id: 'Terra.form.select.selected' });
+    const unselectedTxt = intl.formatMessage({ id: 'Terra.form.select.unselected' });
 
     if (keyCode === KeyCode.KEY_UP) {
       this.clearScrollTimeout();
       this.scrollTimeout = setTimeout(this.clearScrollTimeout, 500);
       this.setState({ active: Util.findPrevious(children, active) });
+      this.updateCurrentActiveScreenReader();
     } else if (keyCode === KeyCode.KEY_DOWN) {
       this.clearScrollTimeout();
       this.scrollTimeout = setTimeout(this.clearScrollTimeout, 500);
       this.setState({ active: Util.findNext(children, active) });
+      this.updateCurrentActiveScreenReader();
     } else if (keyCode === KeyCode.KEY_RETURN && active !== null && (!Util.allowsMultipleSelections(variant) || !Util.includes(value, active))) {
       event.preventDefault();
+      this.setState({ closedViaKeyEvent: true });
       const option = Util.findByValue(children, active);
+      // Handles communicating the case where a clear option is selected to screen readers
+      if (this.props.clearOptionDisplay) {
+        const activeOption = this.menu.querySelector('[data-select-active]');
+        if (activeOption.hasAttribute('data-terra-select-clear-option')) {
+          this.props.visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.selectCleared' });
+        }
+      }
+      // Handles communicating the case where a regular option is selected to screen readers.
+      /*
+        Detecting if browser is not Safari before updating aria-live as there is some odd behaivor
+        with VoiceOver on desktop, that causes the selected option to be read twice when this is
+        is added to aria-live container.
+        When we shift focus back to select, VoiceOver automatically reads the display text.
+        Using aria-hidden on the display does not prevent VO from reading the display text and so it
+        results in reading the display text followed by reading the aria-live message which is
+        the display text + 'selected'
+        */
+      if (SharedUtil.isSafari()) {
+        if (variant === Variants.MULTIPLE || variant === Variants.TAG) {
+          this.props.visuallyHiddenComponent.current.innerText = `${option.props.display} ${selectedTxt}`;
+        }
+      } else {
+        this.props.visuallyHiddenComponent.current.innerText = `${option.props.display} ${selectedTxt}`;
+      }
       onSelect(option.props.value, option);
+    } else if (keyCode === KeyCode.KEY_RETURN && active !== null && Util.allowsMultipleSelections(variant) && Util.includes(value, active)) {
+      event.preventDefault();
+      const option = Util.findByValue(children, active);
+      // Handles communicating the case where a regular option is Unselected to screen readers.
+      /*
+        Detecting if browser is not Safari before updating aria-live as there is some odd behaivor
+        with VoiceOver on desktop, that causes the selected option to be read twice when this is
+        is added to aria-live container.
+        When we shift focus back to select, VoiceOver automatically reads the display text.
+        Using aria-hidden on the display does not prevent VO from reading the display text and so it
+        results in reading the display text followed by reading the aria-live message which is
+        the display text + 'Unselected'
+        */
+      if (SharedUtil.isSafari()) {
+        if (variant === Variants.MULTIPLE || variant === Variants.TAG) {
+          this.props.visuallyHiddenComponent.current.innerText = `${option.props.display} ${unselectedTxt}`;
+        }
+      } else {
+        this.props.visuallyHiddenComponent.current.innerText = `${option.props.display} ${unselectedTxt}`;
+      }
+      onDeselect(option.props.value, option);
     } else if (keyCode === KeyCode.KEY_HOME) {
       event.preventDefault();
       this.setState({ active: Util.findFirst(children) });
@@ -289,20 +448,25 @@ class Menu extends React.Component {
     }
 
     const {
-      onDeselect, onSelect, value, variant,
+      select, input, onDeselect, onSelect, value, variant,
     } = this.props;
 
     if (Util.allowsMultipleSelections(variant) && Util.includes(value, option.props.value)) {
       onDeselect(option.props.value, option);
     } else {
       onSelect(option.props.value, option);
+      if (variant === Variants.DEFAULT) {
+        select.focus();
+      } else {
+        input.focus();
+      }
     }
   }
 
   /**
    * Sets the hovered option as the active value.
    * @param {event} event - The mouse enter event.
-   * @param {ReactNode} option - The option that recieved the mouse enter event.
+   * @param {ReactNode} option - The option that received the mouse enter event.
    */
   handleMouseEnter(event, option) {
     // Prevents setting the active option on mouse enter if the keyboard scrolled the view.
@@ -343,19 +507,23 @@ class Menu extends React.Component {
 
   render() {
     return (
-      // This warns that aria-activedescendant should map to an id
-      // Our implementation maps it to a dynamic id but linter is unable to detect mapping and throws an error
-      /* eslint-disable jsx-a11y/aria-proptypes */
+      /**
+       * Note: role="listbox" and aria-activedescendant needed for VoiceOver on iOS to properly
+       * shift the virtual indicator to this DOM element when dropdown is rendered. If you modify these
+       * attributes, you'll need to manually verify that the virtual indicator on iOS is still shifted
+       * to the dropdown / the selected item in the dropdown if an item is selected when the dropdown
+       * is opened.
+       */
       <ul
+        id="terra-select-menu"
         role="listbox"
         className={cx('menu')}
         ref={(menu) => { this.menu = menu; }}
-        aria-activedescendant={`terra-select-option-${this.state.active}`}
+        {...(this.state.active !== null ? { 'aria-activedescendant': `terra-select-option-${this.state.active}` } : {})}
         tabIndex="0"
       >
         {this.clone(this.state.children)}
       </ul>
-      /* eslint-enable jsx-a11y/aria-proptypes */
     );
   }
 }
@@ -365,9 +533,9 @@ Menu.defaultProps = defaultProps;
 Menu.contextTypes = contextTypes;
 
 /**
- * This polyfill enables backwards compatability of features added in React 16.3.0.
+ * This polyfill enables backwards compatibility of features added in React 16.3.0.
  * More information is available at: https://github.com/reactjs/react-lifecycles-compat
  */
 polyfill(Menu);
 
-export default Menu;
+export default injectIntl(Menu);