+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+5.11.0 - (March 26, 2019)
+------------------
 ### Changed
 * Update react/react-dom peer-dependency to latest stable minor release
 
@@ -35,6 +38,7 @@ Unreleased
 ------------------
 ### Fixed
 * Race condition where timeout would update after the component is not mounted
+* Improved experience using the select component with JAWS, VoiceOver, and VoiceOver on iOS
 * 'Unable to get property 'getInstance' of undefined or null reference' error in IE
 
 5.7.0 - (February 26, 2019)